refactor NameForm
