+++ src/pages/floor/[index].js
@@ -70,7 +70,7 @@ const StyledSection = styled.section`
   left: 50%;
   transform: translateX(-50%);
   height: 527px;
-  width: 100%;
+  width: 357px;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -78,13 +78,18 @@ const StyledSection = styled.section`
   border: 2px solid red;
   transition: top 0.5s ease-in-out;
   overflow: scroll;
+  scrollbar-width: none;
 
   @media (min-height: 740px) {
     height: 600px;
   }
 
   @media (min-height: 770px) {
-    height: 630px;
+    height: 700px;
+  }
+
+  ::-webkit-scrollbar {
+    display: none;
   }
 `;
