+++ styles.js
@@ -77,7 +77,7 @@ export default createGlobalStyle`
     margin: 0;
     background-color: var(--6);
     color: var(--1);
-    overflow: hidden;
+    
   }
 
   input,
